+++ sdk/ai/azure-ai-resources/azure/ai/resources/client/_user_agent.py
@@ -1,0 +1,6 @@
+# ---------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# ---------------------------------------------------------
+from azure.ai.resources._version import VERSION
+
+USER_AGENT = "{}/{}".format("azure-ai-resources", VERSION)