+++ sdk/contentsafety/azure-ai-contentsafety/azure/ai/contentsafety/_operations/_operations.py
@@ -22,15 +22,14 @@ from azure.core.exceptions import (
 )
 from azure.core.paging import ItemPaged
 from azure.core.pipeline import PipelineResponse
-from azure.core.pipeline.transport import HttpResponse
-from azure.core.rest import HttpRequest
+from azure.core.rest import HttpRequest, HttpResponse
 from azure.core.tracing.decorator import distributed_trace
 from azure.core.utils import case_insensitive_dict
 
 from .. import models as _models
 from .._model_base import AzureJSONEncoder, _deserialize
 from .._serialization import Serializer
-from .._vendor import ContentSafetyClientMixinABC, _format_url_section
+from .._vendor import ContentSafetyClientMixinABC
 
 if sys.version_info >= (3, 9):
     from collections.abc import MutableMapping
@@ -88,7 +87,9 @@ def build_content_safety_analyze_image_request(**kwargs: Any) -> HttpRequest:  #
     return HttpRequest(method="POST", url=_url, params=_params, headers=_headers, **kwargs)
 
 
-def build_content_safety_get_text_blocklist_request(blocklist_name: str, **kwargs: Any) -> HttpRequest:  # pylint: disable=name-too-long
+def build_content_safety_get_text_blocklist_request(  # pylint: disable=name-too-long
+    blocklist_name: str, **kwargs: Any
+) -> HttpRequest:
     _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
     _params = case_insensitive_dict(kwargs.pop("params", {}) or {})
 
@@ -101,7 +102,7 @@ def build_content_safety_get_text_blocklist_request(blocklist_name: str, **kwarg
         "blocklistName": _SERIALIZER.url("blocklist_name", blocklist_name, "str"),
     }
 
-    _url: str = _format_url_section(_url, **path_format_arguments)  # type: ignore
+    _url: str = _url.format(**path_format_arguments)  # type: ignore
 
     # Construct parameters
     _params["api-version"] = _SERIALIZER.query("api_version", api_version, "str")
@@ -112,7 +113,9 @@ def build_content_safety_get_text_blocklist_request(blocklist_name: str, **kwarg
     return HttpRequest(method="GET", url=_url, params=_params, headers=_headers, **kwargs)
 
 
-def build_content_safety_create_or_update_text_blocklist_request(blocklist_name: str, **kwargs: Any) -> HttpRequest:  # pylint: disable=name-too-long
+def build_content_safety_create_or_update_text_blocklist_request(  # pylint: disable=name-too-long
+    blocklist_name: str, **kwargs: Any
+) -> HttpRequest:
     _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
     _params = case_insensitive_dict(kwargs.pop("params", {}) or {})
 
@@ -126,7 +129,7 @@ def build_content_safety_create_or_update_text_blocklist_request(blocklist_name:
         "blocklistName": _SERIALIZER.url("blocklist_name", blocklist_name, "str"),
     }
 
-    _url: str = _format_url_section(_url, **path_format_arguments)  # type: ignore
+    _url: str = _url.format(**path_format_arguments)  # type: ignore
 
     # Construct parameters
     _params["api-version"] = _SERIALIZER.query("api_version", api_version, "str")
@@ -139,7 +142,9 @@ def build_content_safety_create_or_update_text_blocklist_request(blocklist_name:
     return HttpRequest(method="PATCH", url=_url, params=_params, headers=_headers, **kwargs)
 
 
-def build_content_safety_delete_text_blocklist_request(blocklist_name: str, **kwargs: Any) -> HttpRequest:  # pylint: disable=name-too-long
+def build_content_safety_delete_text_blocklist_request(  # pylint: disable=name-too-long
+    blocklist_name: str, **kwargs: Any
+) -> HttpRequest:
     _params = case_insensitive_dict(kwargs.pop("params", {}) or {})
 
     api_version: str = kwargs.pop("api_version", _params.pop("api-version", "2023-04-30-preview"))
@@ -149,7 +154,7 @@ def build_content_safety_delete_text_blocklist_request(blocklist_name: str, **kw
         "blocklistName": _SERIALIZER.url("blocklist_name", blocklist_name, "str"),
     }
 
-    _url: str = _format_url_section(_url, **path_format_arguments)  # type: ignore
+    _url: str = _url.format(**path_format_arguments)  # type: ignore
 
     # Construct parameters
     _params["api-version"] = _SERIALIZER.query("api_version", api_version, "str")
@@ -176,7 +181,9 @@ def build_content_safety_list_text_blocklists_request(**kwargs: Any) -> HttpRequ
     return HttpRequest(method="GET", url=_url, params=_params, headers=_headers, **kwargs)
 
 
-def build_content_safety_add_block_items_request(blocklist_name: str, **kwargs: Any) -> HttpRequest:  # pylint: disable=name-too-long
+def build_content_safety_add_block_items_request(  # pylint: disable=name-too-long
+    blocklist_name: str, **kwargs: Any
+) -> HttpRequest:
     _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
     _params = case_insensitive_dict(kwargs.pop("params", {}) or {})
 
@@ -190,7 +197,7 @@ def build_content_safety_add_block_items_request(blocklist_name: str, **kwargs: 
         "blocklistName": _SERIALIZER.url("blocklist_name", blocklist_name, "str"),
     }
 
-    _url: str = _format_url_section(_url, **path_format_arguments)  # type: ignore
+    _url: str = _url.format(**path_format_arguments)  # type: ignore
 
     # Construct parameters
     _params["api-version"] = _SERIALIZER.query("api_version", api_version, "str")
@@ -203,7 +210,9 @@ def build_content_safety_add_block_items_request(blocklist_name: str, **kwargs: 
     return HttpRequest(method="POST", url=_url, params=_params, headers=_headers, **kwargs)
 
 
-def build_content_safety_remove_block_items_request(blocklist_name: str, **kwargs: Any) -> HttpRequest:  # pylint: disable=name-too-long
+def build_content_safety_remove_block_items_request(  # pylint: disable=name-too-long
+    blocklist_name: str, **kwargs: Any
+) -> HttpRequest:
     _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
     _params = case_insensitive_dict(kwargs.pop("params", {}) or {})
 
@@ -215,7 +224,7 @@ def build_content_safety_remove_block_items_request(blocklist_name: str, **kwarg
         "blocklistName": _SERIALIZER.url("blocklist_name", blocklist_name, "str"),
     }
 
-    _url: str = _format_url_section(_url, **path_format_arguments)  # type: ignore
+    _url: str = _url.format(**path_format_arguments)  # type: ignore
 
     # Construct parameters
     _params["api-version"] = _SERIALIZER.query("api_version", api_version, "str")
@@ -243,7 +252,7 @@ def build_content_safety_get_text_blocklist_item_request(  # pylint: disable=nam
         "blockItemId": _SERIALIZER.url("block_item_id", block_item_id, "str"),
     }
 
-    _url: str = _format_url_section(_url, **path_format_arguments)  # type: ignore
+    _url: str = _url.format(**path_format_arguments)  # type: ignore
 
     # Construct parameters
     _params["api-version"] = _SERIALIZER.query("api_version", api_version, "str")
@@ -255,7 +264,12 @@ def build_content_safety_get_text_blocklist_item_request(  # pylint: disable=nam
 
 
 def build_content_safety_list_text_blocklist_items_request(  # pylint: disable=name-too-long
-    blocklist_name: str, *, top: Optional[int] = None, skip: Optional[int] = None, **kwargs: Any
+    blocklist_name: str,
+    *,
+    top: Optional[int] = None,
+    skip: Optional[int] = None,
+    maxpagesize: Optional[int] = None,
+    **kwargs: Any
 ) -> HttpRequest:
     _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
     _params = case_insensitive_dict(kwargs.pop("params", {}) or {})
@@ -269,7 +283,7 @@ def build_content_safety_list_text_blocklist_items_request(  # pylint: disable=n
         "blocklistName": _SERIALIZER.url("blocklist_name", blocklist_name, "str"),
     }
 
-    _url: str = _format_url_section(_url, **path_format_arguments)  # type: ignore
+    _url: str = _url.format(**path_format_arguments)  # type: ignore
 
     # Construct parameters
     _params["api-version"] = _SERIALIZER.query("api_version", api_version, "str")
@@ -277,6 +291,8 @@ def build_content_safety_list_text_blocklist_items_request(  # pylint: disable=n
         _params["top"] = _SERIALIZER.query("top", top, "int")
     if skip is not None:
         _params["skip"] = _SERIALIZER.query("skip", skip, "int")
+    if maxpagesize is not None:
+        _params["maxpagesize"] = _SERIALIZER.query("maxpagesize", maxpagesize, "int")
 
     # Construct headers
     _headers["Accept"] = _SERIALIZER.header("accept", accept, "str")
@@ -388,9 +404,9 @@ class ContentSafetyClientOperationsMixin(ContentSafetyClientMixinABC):
         if isinstance(body, (IOBase, bytes)):
             _content = body
         else:
-            _content = json.dumps(body, cls=AzureJSONEncoder)  # type: ignore
+            _content = json.dumps(body, cls=AzureJSONEncoder, exclude_readonly=True)  # type: ignore
 
-        request = build_content_safety_analyze_text_request(
+        _request = build_content_safety_analyze_text_request(
             content_type=content_type,
             api_version=self._config.api_version,
             content=_content,
@@ -400,16 +416,18 @@ class ContentSafetyClientOperationsMixin(ContentSafetyClientMixinABC):
         path_format_arguments = {
             "endpoint": self._serialize.url("self._config.endpoint", self._config.endpoint, "str", skip_quote=True),
         }
-        request.url = self._client.format_url(request.url, **path_format_arguments)
+        _request.url = self._client.format_url(_request.url, **path_format_arguments)
 
         _stream = kwargs.pop("stream", False)
         pipeline_response: PipelineResponse = self._client._pipeline.run(  # pylint: disable=protected-access
-            request, stream=_stream, **kwargs
+            _request, stream=_stream, **kwargs
         )
 
         response = pipeline_response.http_response
 
         if response.status_code not in [200]:
+            if _stream:
+                response.read()  # Load the body in memory and close the socket
             map_error(status_code=response.status_code, response=response, error_map=error_map)
             raise HttpResponseError(response=response)
 
@@ -526,9 +544,9 @@ class ContentSafetyClientOperationsMixin(ContentSafetyClientMixinABC):
         if isinstance(body, (IOBase, bytes)):
             _content = body
         else:
-            _content = json.dumps(body, cls=AzureJSONEncoder)  # type: ignore
+            _content = json.dumps(body, cls=AzureJSONEncoder, exclude_readonly=True)  # type: ignore
 
-        request = build_content_safety_analyze_image_request(
+        _request = build_content_safety_analyze_image_request(
             content_type=content_type,
             api_version=self._config.api_version,
             content=_content,
@@ -538,16 +556,18 @@ class ContentSafetyClientOperationsMixin(ContentSafetyClientMixinABC):
         path_format_arguments = {
             "endpoint": self._serialize.url("self._config.endpoint", self._config.endpoint, "str", skip_quote=True),
         }
-        request.url = self._client.format_url(request.url, **path_format_arguments)
+        _request.url = self._client.format_url(_request.url, **path_format_arguments)
 
         _stream = kwargs.pop("stream", False)
         pipeline_response: PipelineResponse = self._client._pipeline.run(  # pylint: disable=protected-access
-            request, stream=_stream, **kwargs
+            _request, stream=_stream, **kwargs
         )
 
         response = pipeline_response.http_response
 
         if response.status_code not in [200]:
+            if _stream:
+                response.read()  # Load the body in memory and close the socket
             map_error(status_code=response.status_code, response=response, error_map=error_map)
             raise HttpResponseError(response=response)
 
@@ -588,7 +608,7 @@ class ContentSafetyClientOperationsMixin(ContentSafetyClientMixinABC):
 
         cls: ClsType[_models.TextBlocklist] = kwargs.pop("cls", None)
 
-        request = build_content_safety_get_text_blocklist_request(
+        _request = build_content_safety_get_text_blocklist_request(
             blocklist_name=blocklist_name,
             api_version=self._config.api_version,
             headers=_headers,
@@ -597,16 +617,18 @@ class ContentSafetyClientOperationsMixin(ContentSafetyClientMixinABC):
         path_format_arguments = {
             "endpoint": self._serialize.url("self._config.endpoint", self._config.endpoint, "str", skip_quote=True),
         }
-        request.url = self._client.format_url(request.url, **path_format_arguments)
+        _request.url = self._client.format_url(_request.url, **path_format_arguments)
 
         _stream = kwargs.pop("stream", False)
         pipeline_response: PipelineResponse = self._client._pipeline.run(  # pylint: disable=protected-access
-            request, stream=_stream, **kwargs
+            _request, stream=_stream, **kwargs
         )
 
         response = pipeline_response.http_response
 
         if response.status_code not in [200]:
+            if _stream:
+                response.read()  # Load the body in memory and close the socket
             map_error(status_code=response.status_code, response=response, error_map=error_map)
             raise HttpResponseError(response=response)
 
@@ -731,9 +753,9 @@ class ContentSafetyClientOperationsMixin(ContentSafetyClientMixinABC):
         if isinstance(resource, (IOBase, bytes)):
             _content = resource
         else:
-            _content = json.dumps(resource, cls=AzureJSONEncoder)  # type: ignore
+            _content = json.dumps(resource, cls=AzureJSONEncoder, exclude_readonly=True)  # type: ignore
 
-        request = build_content_safety_create_or_update_text_blocklist_request(
+        _request = build_content_safety_create_or_update_text_blocklist_request(
             blocklist_name=blocklist_name,
             content_type=content_type,
             api_version=self._config.api_version,
@@ -744,16 +766,18 @@ class ContentSafetyClientOperationsMixin(ContentSafetyClientMixinABC):
         path_format_arguments = {
             "endpoint": self._serialize.url("self._config.endpoint", self._config.endpoint, "str", skip_quote=True),
         }
-        request.url = self._client.format_url(request.url, **path_format_arguments)
+        _request.url = self._client.format_url(_request.url, **path_format_arguments)
 
         _stream = kwargs.pop("stream", False)
         pipeline_response: PipelineResponse = self._client._pipeline.run(  # pylint: disable=protected-access
-            request, stream=_stream, **kwargs
+            _request, stream=_stream, **kwargs
         )
 
         response = pipeline_response.http_response
 
         if response.status_code not in [200, 201]:
+            if _stream:
+                response.read()  # Load the body in memory and close the socket
             map_error(status_code=response.status_code, response=response, error_map=error_map)
             raise HttpResponseError(response=response)
 
@@ -803,7 +827,7 @@ class ContentSafetyClientOperationsMixin(ContentSafetyClientMixinABC):
 
         cls: ClsType[None] = kwargs.pop("cls", None)
 
-        request = build_content_safety_delete_text_blocklist_request(
+        _request = build_content_safety_delete_text_blocklist_request(
             blocklist_name=blocklist_name,
             api_version=self._config.api_version,
             headers=_headers,
@@ -812,21 +836,23 @@ class ContentSafetyClientOperationsMixin(ContentSafetyClientMixinABC):
         path_format_arguments = {
             "endpoint": self._serialize.url("self._config.endpoint", self._config.endpoint, "str", skip_quote=True),
         }
-        request.url = self._client.format_url(request.url, **path_format_arguments)
+        _request.url = self._client.format_url(_request.url, **path_format_arguments)
 
         _stream = kwargs.pop("stream", False)
         pipeline_response: PipelineResponse = self._client._pipeline.run(  # pylint: disable=protected-access
-            request, stream=_stream, **kwargs
+            _request, stream=_stream, **kwargs
         )
 
         response = pipeline_response.http_response
 
         if response.status_code not in [204]:
+            if _stream:
+                response.read()  # Load the body in memory and close the socket
             map_error(status_code=response.status_code, response=response, error_map=error_map)
             raise HttpResponseError(response=response)
 
         if cls:
-            return cls(pipeline_response, None, {})
+            return cls(pipeline_response, None, {})  # type: ignore
 
     @distributed_trace
     def list_text_blocklists(self, **kwargs: Any) -> Iterable["_models.TextBlocklist"]:
@@ -854,7 +880,7 @@ class ContentSafetyClientOperationsMixin(ContentSafetyClientMixinABC):
         def prepare_request(next_link=None):
             if not next_link:
 
-                request = build_content_safety_list_text_blocklists_request(
+                _request = build_content_safety_list_text_blocklists_request(
                     api_version=self._config.api_version,
                     headers=_headers,
                     params=_params,
@@ -864,7 +890,7 @@ class ContentSafetyClientOperationsMixin(ContentSafetyClientMixinABC):
                         "self._config.endpoint", self._config.endpoint, "str", skip_quote=True
                     ),
                 }
-                request.url = self._client.format_url(request.url, **path_format_arguments)
+                _request.url = self._client.format_url(_request.url, **path_format_arguments)
 
             else:
                 # make call to next link with the client's api-version
@@ -876,7 +902,7 @@ class ContentSafetyClientOperationsMixin(ContentSafetyClientMixinABC):
                     }
                 )
                 _next_request_params["api-version"] = self._config.api_version
-                request = HttpRequest(
+                _request = HttpRequest(
                     "GET", urllib.parse.urljoin(next_link, _parsed_next_link.path), params=_next_request_params
                 )
                 path_format_arguments = {
@@ -884,9 +910,9 @@ class ContentSafetyClientOperationsMixin(ContentSafetyClientMixinABC):
                         "self._config.endpoint", self._config.endpoint, "str", skip_quote=True
                     ),
                 }
-                request.url = self._client.format_url(request.url, **path_format_arguments)
+                _request.url = self._client.format_url(_request.url, **path_format_arguments)
 
-            return request
+            return _request
 
         def extract_data(pipeline_response):
             deserialized = pipeline_response.http_response.json()
@@ -896,15 +922,17 @@ class ContentSafetyClientOperationsMixin(ContentSafetyClientMixinABC):
             return deserialized.get("nextLink") or None, iter(list_of_elem)
 
         def get_next(next_link=None):
-            request = prepare_request(next_link)
+            _request = prepare_request(next_link)
 
             _stream = False
             pipeline_response: PipelineResponse = self._client._pipeline.run(  # pylint: disable=protected-access
-                request, stream=_stream, **kwargs
+                _request, stream=_stream, **kwargs
             )
             response = pipeline_response.http_response
 
             if response.status_code not in [200]:
+                if _stream:
+                    response.read()  # Load the body in memory and close the socket
                 map_error(status_code=response.status_code, response=response, error_map=error_map)
                 raise HttpResponseError(response=response)
 
@@ -1023,9 +1051,9 @@ class ContentSafetyClientOperationsMixin(ContentSafetyClientMixinABC):
         if isinstance(body, (IOBase, bytes)):
             _content = body
         else:
-            _content = json.dumps(body, cls=AzureJSONEncoder)  # type: ignore
+            _content = json.dumps(body, cls=AzureJSONEncoder, exclude_readonly=True)  # type: ignore
 
-        request = build_content_safety_add_block_items_request(
+        _request = build_content_safety_add_block_items_request(
             blocklist_name=blocklist_name,
             content_type=content_type,
             api_version=self._config.api_version,
@@ -1036,16 +1064,18 @@ class ContentSafetyClientOperationsMixin(ContentSafetyClientMixinABC):
         path_format_arguments = {
             "endpoint": self._serialize.url("self._config.endpoint", self._config.endpoint, "str", skip_quote=True),
         }
-        request.url = self._client.format_url(request.url, **path_format_arguments)
+        _request.url = self._client.format_url(_request.url, **path_format_arguments)
 
         _stream = kwargs.pop("stream", False)
         pipeline_response: PipelineResponse = self._client._pipeline.run(  # pylint: disable=protected-access
-            request, stream=_stream, **kwargs
+            _request, stream=_stream, **kwargs
         )
 
         response = pipeline_response.http_response
 
         if response.status_code not in [200]:
+            if _stream:
+                response.read()  # Load the body in memory and close the socket
             map_error(status_code=response.status_code, response=response, error_map=error_map)
             raise HttpResponseError(response=response)
 
@@ -1170,9 +1200,9 @@ class ContentSafetyClientOperationsMixin(ContentSafetyClientMixinABC):
         if isinstance(body, (IOBase, bytes)):
             _content = body
         else:
-            _content = json.dumps(body, cls=AzureJSONEncoder)  # type: ignore
+            _content = json.dumps(body, cls=AzureJSONEncoder, exclude_readonly=True)  # type: ignore
 
-        request = build_content_safety_remove_block_items_request(
+        _request = build_content_safety_remove_block_items_request(
             blocklist_name=blocklist_name,
             content_type=content_type,
             api_version=self._config.api_version,
@@ -1183,21 +1213,23 @@ class ContentSafetyClientOperationsMixin(ContentSafetyClientMixinABC):
         path_format_arguments = {
             "endpoint": self._serialize.url("self._config.endpoint", self._config.endpoint, "str", skip_quote=True),
         }
-        request.url = self._client.format_url(request.url, **path_format_arguments)
+        _request.url = self._client.format_url(_request.url, **path_format_arguments)
 
         _stream = kwargs.pop("stream", False)
         pipeline_response: PipelineResponse = self._client._pipeline.run(  # pylint: disable=protected-access
-            request, stream=_stream, **kwargs
+            _request, stream=_stream, **kwargs
         )
 
         response = pipeline_response.http_response
 
         if response.status_code not in [204]:
+            if _stream:
+                response.read()  # Load the body in memory and close the socket
             map_error(status_code=response.status_code, response=response, error_map=error_map)
             raise HttpResponseError(response=response)
 
         if cls:
-            return cls(pipeline_response, None, {})
+            return cls(pipeline_response, None, {})  # type: ignore
 
     @distributed_trace
     def get_text_blocklist_item(self, blocklist_name: str, block_item_id: str, **kwargs: Any) -> _models.TextBlockItem:
@@ -1228,7 +1260,7 @@ class ContentSafetyClientOperationsMixin(ContentSafetyClientMixinABC):
 
         cls: ClsType[_models.TextBlockItem] = kwargs.pop("cls", None)
 
-        request = build_content_safety_get_text_blocklist_item_request(
+        _request = build_content_safety_get_text_blocklist_item_request(
             blocklist_name=blocklist_name,
             block_item_id=block_item_id,
             api_version=self._config.api_version,
@@ -1238,16 +1270,18 @@ class ContentSafetyClientOperationsMixin(ContentSafetyClientMixinABC):
         path_format_arguments = {
             "endpoint": self._serialize.url("self._config.endpoint", self._config.endpoint, "str", skip_quote=True),
         }
-        request.url = self._client.format_url(request.url, **path_format_arguments)
+        _request.url = self._client.format_url(_request.url, **path_format_arguments)
 
         _stream = kwargs.pop("stream", False)
         pipeline_response: PipelineResponse = self._client._pipeline.run(  # pylint: disable=protected-access
-            request, stream=_stream, **kwargs
+            _request, stream=_stream, **kwargs
         )
 
         response = pipeline_response.http_response
 
         if response.status_code not in [200]:
+            if _stream:
+                response.read()  # Load the body in memory and close the socket
             map_error(status_code=response.status_code, response=response, error_map=error_map)
             raise HttpResponseError(response=response)
 
@@ -1282,6 +1316,7 @@ class ContentSafetyClientOperationsMixin(ContentSafetyClientMixinABC):
         _headers = kwargs.pop("headers", {}) or {}
         _params = kwargs.pop("params", {}) or {}
 
+        maxpagesize = kwargs.pop("maxpagesize", None)
         cls: ClsType[List[_models.TextBlockItem]] = kwargs.pop("cls", None)
 
         error_map = {
@@ -1295,10 +1330,11 @@ class ContentSafetyClientOperationsMixin(ContentSafetyClientMixinABC):
         def prepare_request(next_link=None):
             if not next_link:
 
-                request = build_content_safety_list_text_blocklist_items_request(
+                _request = build_content_safety_list_text_blocklist_items_request(
                     blocklist_name=blocklist_name,
                     top=top,
                     skip=skip,
+                    maxpagesize=maxpagesize,
                     api_version=self._config.api_version,
                     headers=_headers,
                     params=_params,
@@ -1308,7 +1344,7 @@ class ContentSafetyClientOperationsMixin(ContentSafetyClientMixinABC):
                         "self._config.endpoint", self._config.endpoint, "str", skip_quote=True
                     ),
                 }
-                request.url = self._client.format_url(request.url, **path_format_arguments)
+                _request.url = self._client.format_url(_request.url, **path_format_arguments)
 
             else:
                 # make call to next link with the client's api-version
@@ -1320,7 +1356,7 @@ class ContentSafetyClientOperationsMixin(ContentSafetyClientMixinABC):
                     }
                 )
                 _next_request_params["api-version"] = self._config.api_version
-                request = HttpRequest(
+                _request = HttpRequest(
                     "GET", urllib.parse.urljoin(next_link, _parsed_next_link.path), params=_next_request_params
                 )
                 path_format_arguments = {
@@ -1328,9 +1364,9 @@ class ContentSafetyClientOperationsMixin(ContentSafetyClientMixinABC):
                         "self._config.endpoint", self._config.endpoint, "str", skip_quote=True
                     ),
                 }
-                request.url = self._client.format_url(request.url, **path_format_arguments)
+                _request.url = self._client.format_url(_request.url, **path_format_arguments)
 
-            return request
+            return _request
 
         def extract_data(pipeline_response):
             deserialized = pipeline_response.http_response.json()
@@ -1340,15 +1376,17 @@ class ContentSafetyClientOperationsMixin(ContentSafetyClientMixinABC):
             return deserialized.get("nextLink") or None, iter(list_of_elem)
 
         def get_next(next_link=None):
-            request = prepare_request(next_link)
+            _request = prepare_request(next_link)
 
             _stream = False
             pipeline_response: PipelineResponse = self._client._pipeline.run(  # pylint: disable=protected-access
-                request, stream=_stream, **kwargs
+                _request, stream=_stream, **kwargs
             )
             response = pipeline_response.http_response
 
             if response.status_code not in [200]:
+                if _stream:
+                    response.read()  # Load the body in memory and close the socket
                 map_error(status_code=response.status_code, response=response, error_map=error_map)
                 raise HttpResponseError(response=response)