+++ sdk/cosmos/azure-cosmos/samples/examples.py
@@ -109,6 +109,22 @@ for item in discontinued_items:
     print(json.dumps(item, indent=True))
 # [END query_items_param]
 
+# [START priority_level option]
+# Priority-based execution is a capability which allows users to specify priority
+# for the request sent to Azure Cosmos DB. Based on the priority specified by the user,
+# if there are more requests than the configured RU/s in a second,
+# then Azure Cosmos DB will throttle low priority requests to allow high priority requests to execute.
+# Can be used for Read, Write, and Query operations. This is specified with the priority_level keyword.
+# the value can either be low or high.
+for item in container.query_items(
+    query='SELECT * FROM products p WHERE p.productModel <> "DISCONTINUED"',
+    enable_cross_partition_query=True,
+    priority_level="High"
+):
+    print(json.dumps(item, indent=True))
+
+# [END priority_level option]
+
 # Delete items from the container.
 # The Cosmos DB SQL API does not support 'DELETE' queries,
 # so deletes must be done with the delete_item method